requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
argon2-cffi==19.2.0       # via -r requirements.in
asgiref==3.2.7            # via django
cffi==1.14.0              # via argon2-cffi
django-cors-headers==3.3.0  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-extensions==2.2.9  # via -r requirements.in
django-model-utils==4.0.0  # via -r requirements.in
django==3.0.6             # via -r requirements.in, django-cors-headers, django-model-utils, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
pillow==7.1.2             # via -r requirements.in
psycopg2==2.8.5           # via -r requirements.in
pycparser==2.20           # via cffi
pytz==2020.1              # via django
six==1.15.0               # via argon2-cffi, django-extensions
sorl-thumbnail==12.6.3    # via -r requirements.in
sqlparse==0.3.1           # via django
whitenoise==5.0.1         # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
